% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Multi_Gene.R
\name{Get_Multi_Gene}
\alias{Get_Multi_Gene}
\title{Get significant gene list from Multi_Trait Models}
\usage{
Get_Multi_Gene(output, threshold = 0.05)
}
\arguments{
\item{output}{A list object from the output of multi-trait models}

\item{threshold}{Threshold for FDR. Default=0.05}
}
\value{
The significant gene list for two traits under FDR<threshold
\item{Gene_trait1}{Significant gene list for trait 1}
\item{Gene_trait2}{Significant gene list for trait 2}
}
\description{
Get significant gene list from Multi_Trait Models
}
