% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Single_Gene.R
\name{Get_Single_Gene}
\alias{Get_Single_Gene}
\title{Get significant gene list from single-trait models}
\usage{
Get_Single_Gene(output, threshold = 0.05)
}
\arguments{
\item{output}{A list object from the output of single-trait models}

\item{threshold}{Threshold for FDR. Default=0.05}
}
\value{
The significant gene vector for the trait under FDR<threshold
\item{Gene_trait}{Significant gene list for the trait}
}
\description{
Get significant gene list from single-trait models
}
