% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Instruction.R
\docType{data}
\name{Instruction}
\alias{Instruction}
\title{Instruction}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x3647110> 
 - attr(*, "name")= chr "Instruction_generator"
}}
\usage{
Instruction
}
\description{
Instruction
}
\keyword{datasets}

