% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KeyVariables.R
\docType{data}
\name{KEY.VALUES.DATE}
\alias{KEY.VALUES.DATE}
\title{KEY.VALUES.DATE}
\format{\preformatted{ chr "DATE"
}}
\usage{
KEY.VALUES.DATE
}
\description{
KEY.VALUES.DATE
}
\keyword{datasets}

