% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KeyVariables.R
\docType{data}
\name{KEY.VALUES.EFFDATE}
\alias{KEY.VALUES.EFFDATE}
\title{KEY.VALUES.EFFDATE}
\format{\preformatted{ chr "EFFDATE"
}}
\usage{
KEY.VALUES.EFFDATE
}
\description{
KEY.VALUES.EFFDATE
}
\keyword{datasets}

