% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MData.R
\docType{data}
\name{MData}
\alias{MData}
\title{MData}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x28b0a80> 
 - attr(*, "name")= chr "MData_generator"
}}
\usage{
MData
}
\description{
MData
}
\keyword{datasets}

