% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MDataSet.R
\docType{data}
\name{MDataSet}
\alias{MDataSet}
\title{MDataSet}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x33e9d80> 
 - attr(*, "name")= chr "MDataSet_generator"
}}
\usage{
MDataSet
}
\description{
MDataSet
}
\keyword{datasets}

