% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MDataUpdater.R
\docType{data}
\name{MDataUpdater}
\alias{MDataUpdater}
\title{MDataUpdater}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x33954d0> 
 - attr(*, "name")= chr "MDataUpdater_generator"
}}
\usage{
MDataUpdater
}
\description{
MDataUpdater
}
\keyword{datasets}

