% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MParameter.R
\docType{data}
\name{MParameter}
\alias{MParameter}
\title{MParameter}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x2e72c00> 
 - attr(*, "name")= chr "MParameter_generator"
}}
\usage{
MParameter
}
\description{
MParameter
}
\keyword{datasets}

