% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MParameterSet.R
\docType{data}
\name{MParameterSet}
\alias{MParameterSet}
\title{MParameterSet}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x30f2c78> 
 - attr(*, "name")= chr "MParameterSet_generator"
}}
\usage{
MParameterSet
}
\description{
MParameterSet
}
\keyword{datasets}

