% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSet.R
\docType{data}
\name{MSet}
\alias{MSet}
\title{MSet}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x20e5610> 
 - attr(*, "name")= chr "MSet_generator"
}}
\usage{
MSet
}
\description{
MSet
}
\keyword{datasets}

