% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MStack.R
\docType{data}
\name{MStack}
\alias{MStack}
\title{MStack}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x269dd78> 
 - attr(*, "name")= chr "MStack_generator"
}}
\usage{
MStack
}
\description{
MStack
}
\keyword{datasets}

