% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Anno.R
\name{Multi_Anno}
\alias{Multi_Anno}
\title{Run Multi-Trait with Annotation Model}
\usage{
Multi_Anno(
  data,
  Anno1,
  Anno2,
  N_1,
  N_2,
  pi_init = c(0.85, 0.05, 0.05, 0.05),
  beta0_trait1_init,
  beta0_trait2_init,
  threshold_1 = 0.001,
  threshold_2 = 0.001,
  max_iter = 200
)
}
\arguments{
\item{data}{A dataframe with columns: Gene name, mutability, de novo mutation count for trait1, de novo mutation count for trait2}

\item{Anno1}{Annotation file for trait 1}

\item{Anno2}{Annotation file for trait 2}

\item{N_1}{Cohort size for trait 1}

\item{N_2}{Cohort size for trait 2}

\item{pi_init}{Initial value for probabilities of genes under different assumptions. Default=c(0.85,0.05,0.05,0.05)}

\item{beta0_trait1_init}{Initial value for log gamma for trait 1}

\item{beta0_trait2_init}{Initial value for log gamma for trait 2}

\item{threshold_1}{Threshold for EM algorithm. Default=1e-3}

\item{threshold_2}{Threshold for Newton's method. Default=1e-3}

\item{max_iter}{Maximum iteration for Newton's method. Default=200}
}
\value{
The estimated model parameters and the posterior probabilities of genes under different assumptions
\item{result}{A dataframe that includes estimated posterior probabilities of risk genes for each trait and estimated posterior probability for shared risk gene}
\item{pi}{Estimated proportion of genes under different assumptions}
\item{beta_trait1}{Estimated beta vector for trait 1}
\item{beta_trait2}{Estimated beta vector for trait 2}
\item{Z_mat}{Estimated posterior probabilities of genes under different assumptions}
}
\description{
Run Multi-Trait with Annotation Model
}
