% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_No.R
\name{Multi_No}
\alias{Multi_No}
\title{Run Multi-Trait without Annotation Model}
\usage{
Multi_No(
  data,
  N_1,
  N_2,
  pi_init = c(0.85, 0.05, 0.05, 0.05),
  threshold = 1e-06
)
}
\arguments{
\item{data}{A dataframe with columns: Gene name, mutability, de novo mutation count for trait1, de novo mutation count for trait2}

\item{N_1}{Cohort size for trait 1}

\item{N_2}{Cohort size for trait 2}

\item{pi_init}{Initial value for probabilities of genes under different assumptions. Default=c(0.85,0.05,0.05,0.05)}

\item{threshold}{Threshold for EM algorithm. Default=1e-6}
}
\value{
The estimated model parameters and the posterior probabilities of genes under different assumptions
\item{result}{A dataframe that includes estimated posterior probabilities of risk genes for each trait and estimated posterior probability for shared risk gene}
\item{pi}{Estimated proportion of genes under different assumptions}
\item{beta_trait1}{Estimated log of gamma for trait 1}
\item{beta_trait2}{Estimated log of gamma for trait 2}
\item{Z_mat}{Estimated posterior probabilities of genes under different assumptions}
}
\description{
Run Multi-Trait without Annotation Model
}
