% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Anno.R
\name{Single_Anno}
\alias{Single_Anno}
\title{Run Single-Trait with Annotation Model}
\usage{
Single_Anno(
  data,
  dn_col,
  N_1,
  Anno,
  pi_init = 0.1,
  beta0_init,
  threshold_1 = 0.001,
  threshold_2 = 0.001,
  max_iter = 200
)
}
\arguments{
\item{data}{A dataframe with columns: Gene name, mutability, de novo mutation count for traits}

\item{dn_col}{A character indicating the column name of de novo mutation counts for the trait}

\item{N_1}{Cohort size for the trait}

\item{Anno}{Annotation file for the trait}

\item{pi_init}{Initial value for risk gene proportion. Default=0.1}

\item{threshold_1}{Threshold for EM algorithm. Default=1e-3}

\item{threshold_2}{Threshold for Newton's method. Default=1e-3}

\item{max_iter}{Maximum iteration for Newton's method. Default=200}
}
\value{
The estimated model parameters and the posterior probabilities of genes under different assumptions
\item{result}{A dataframe that includes estimated posterior probabilities of risk genes for each trait and estimated posterior probability for shared risk gene}
\item{pi}{Estimated proportion vector, the second value represents risk gene proportion}
\item{beta}{Estimated beta vector}
\item{Z_mat}{Estimated posterior probabilities of genes under different assumptions}
}
\description{
Run Single-Trait with Annotation Model
}
