% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_No.R
\name{Single_No}
\alias{Single_No}
\title{Run Single-Trait without Annotation Model}
\usage{
Single_No(data, dn_col, N_1, pi_init = 0.1, threshold = 1e-06)
}
\arguments{
\item{data}{A dataframe with columns: Gene name, mutability, de novo mutation count for traits}

\item{dn_col}{A character indicating the column name of de novo mutation counts for the trait}

\item{N_1}{Cohort size for the trait}

\item{pi_init}{Initial value for risk gene proportion. Default=0.1}

\item{threshold}{Threshold for EM algorithm. Default=1e-6}
}
\value{
The estimated model parameters and the posterior probabilities of genes under different assumptions
\item{result}{A dataframe that includes estimated posterior probabilities of risk genes for each trait and estimated posterior probability for shared risk gene}
\item{pi}{Estimated proportion vector, the second value represents risk gene proportion}
\item{beta}{Estimated log of gamma for risk genes}
\item{Z_mat}{Estimated posterior probabilities of genes under different assumptions}
}
\description{
Run Single-Trait without Annotation Model
}
