% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UpdaterSet.R
\docType{data}
\name{UpdaterSet}
\alias{UpdaterSet}
\title{UpdaterSet}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x26d65f8> 
 - attr(*, "name")= chr "UpdaterSet_generator"
}}
\usage{
UpdaterSet
}
\description{
UpdaterSet
}
\keyword{datasets}

