% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MDataManager.R
\docType{data}
\name{MDataManager}
\alias{MDataManager}
\title{UpdaterSet}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x35f5da0> 
 - attr(*, "name")= chr "MDataManager_generator"
}}
\usage{
MDataManager
}
\description{
UpdaterSet
}
\keyword{datasets}

