% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MInstructionStack.R
\docType{data}
\name{MInstructionStack}
\alias{MInstructionStack}
\title{MStack}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x335f8e8> 
 - attr(*, "name")= chr "MInstructionStack_generator"
}}
\usage{
MInstructionStack
}
\description{
MStack
}
\keyword{datasets}

